agent/requirements.txt
# Core Flask dependencies
Flask==2.3.3
Flask-Bootstrap==3.3.7.1
Flask-WTF==1.2.1
WTForms==3.1.0
email-validator==2.1.0

# Database dependencies
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Security dependencies
bcrypt==4.1.2
python-dotenv==1.0.0

# Existing project dependencies
opencv-python==4.8.1.78
ultralytics==8.0.196
torch==2.1.1
torchvision==0.16.1
numpy==1.24.3
Pillow==10.1.0
scipy==1.11.4
matplotlib==3.8.2

# Additional utilities
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.7.0